algorithm: change early exit to a function call
We change the early exit to `goal:found(current)` to allow for different
node objects to determine when the goal has been met. This allows
for more varied types of graphs, like a goal oriented action plan
(GOAP) graph.

Also fix a luacheck error with current not being defined at the correct
scope.