README: update environment setup instructions
Convert to setting up using Window's 'wsl' environment for development.
This simplifies the process and doesn't leave people questioning what
to do if they are not really familiar with Linux development
environments.

Also, updated the copyright years in the library lua. Updated
devel-setup to reflect the current development dependencies and
cleaned up the Makefile.