class: add the concept of properties
This allows for automatic getter/setter behaviour.
Example:

local A = class("A")
function A:__init()
	self:_property("foo", 2,
		function(self, key, new, old)
			local v = tonumber(new)
			if v == nil then
				return old
			end
			return v
		end)
end

This will make an instance of A have a property `foo` with a
default value of 2 and when attempting to be set will require the
new value to be a number otherwise the previous value will be
kept. This allows accesses like:

local a = A()
print(a.foo)  ---> "2"
a.foo = "dog"
print(a.foo)  ---> "2"
a.foo = 4
print(a.foo)  ---> "4"